全局软删除筛选器 代码调整。
